test: move test_temp_data_dir to store tests
Refs: OPS-70
